fix a lot

+++ install/install-zsh-config.zsh
@@ -13,13 +13,11 @@ if [ -f "~/.zshrc" ]; then
     mv "~/.zshrc" "~/.zshrc-original"
 fi
 
-
 mkdir -p ~/.config
-[ -e "$HOME/.oh-my-zsh" ] || { 
-    ln -s "$HOME/.config/zsh/oh-my-zsh" "$HOME/.oh-my-zsh"
-    sh -c "$(wget -O- https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
-}
+ln -srfT "$INSTALLER_DIR/config/zsh" "$HOME/.config/zsh"
+export ZSH="$HOME/.config/zsh/oh-my-zsh"
+
+[ -e "$ZSH" ] || sh -c "$(wget -O- https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 [ -d "$HOME/.oh-my-zsh/custom/themes/powerlevel10k" ] || git clone --depth=1 https://github.com/romkatv/powerlevel10k.git $HOME/.oh-my-zsh/custom/themes/powerlevel10k
-ln -sfT "$INSTALLER_DIR/config/zsh" "$HOME/.config/zsh"
-ln -sf "$HOME/.config/zsh/zshrc" "$HOME/.zshrc"
-[ -f "$HOME/.p10k.zsh" ] || ln -s "$HOME/.config/zsh/p10k.zsh" "$HOME/.p10k.zsh"
+ln -srf "$HOME/.config/zsh/zshrc" "$HOME/.zshrc"
+[ -f "$HOME/.p10k.zsh" ] || ln -sr "$HOME/.config/zsh/p10k.zsh" "$HOME/.p10k.zsh"
